playerChoices.cpp
This is a basic idea of the combat, the user decides what to do by typing to the terminal (can advance to ncurses when needed) and selecting options. At the moment there is only an attack option which is just a test to make sure it reads the input properly. After this I can add classes somehow to allow more options based on their initial choice. Hope this is a step in the right direction.