A super basic idea of combat with weapon selection
I have some of Matt's code included to start the semi integration process. You won't have to make a character each time but that was just something I need to include in combat for formulas to allow suitable attacks from Matt. Slowly getting closer but another week or so and there'll be multiple options depending on what character class you pick and more random monster to attack. Need a hand with the possible initalisation of combat but that can come later with integration.